requirements.txt
langchain
openai
sentence-transformers
faiss-cpu
pypdf
fastapi
uvicorn
python-dotenv
pandas